Fold fountain retaliation into the anti-camp ability
